add code

+++ ServicePortals.Application/Dtos/Purchase/Requests/CreatePurchaseDetailRequest.cs
@@ -5,7 +5,7 @@
         public Guid? Id { get; set; }
         public string? ItemName { get; set; }
         public string? ItemDescription { get; set; }
-        public int Quantity { get; set; } = 0;
+        public decimal Quantity { get; set; } = 0;
         public string? UnitMeasurement { get; set; }
         public DateTimeOffset? RequiredDate { get; set; }
         public int? CostCenterId { get; set; }
